Move getReactionRateDict() into the composites module per reviewer comments
