+++ armi/reactor/components/component.py
@@ -1210,45 +1210,3 @@ class ShapedComponent(Component):
     """A component with well-defined dimensions."""
 
     pass
-
-
-def getReactionRateDict(nucName, lib, xsSuffix, mgFlux, nDens):
-    """
-    Parameters
-    ----------
-    nucName : str
-        nuclide name -- e.g. 'U235', 'PU239', etc. Not to be confused with the nuclide
-        _label_, see the nucDirectory module for a description of the difference.
-    lib : isotxs
-        cross section library
-    xsSuffix : str
-        cross section suffix, consisting of the type followed by the burnup group,
-        e.g. 'AB' for the second burnup group of type A
-    mgFlux : numpy.nArray
-        integrated mgFlux (n-cm/s)
-    nDens : float
-        number density (atom/bn-cm)
-
-    Returns
-    -------
-    rxnRates - dict
-        dictionary of reaction rates (rxn/s) for nG, nF, n2n, nA and nP
-
-    Notes
-    -----
-    Assume there is no n3n cross section in ISOTXS
-    """
-    nucLabel = nuclideBases.byName[nucName].label
-    key = "{}{}".format(nucLabel, xsSuffix)
-    libNuc = lib[key]
-    rxnRates = {"n3n": 0}
-    for rxName, mgXSs in [
-        ("nG", libNuc.micros.nGamma),
-        ("nF", libNuc.micros.fission),
-        ("n2n", libNuc.micros.n2n),
-        ("nA", libNuc.micros.nalph),
-        ("nP", libNuc.micros.np),
-    ]:
-        rxnRates[rxName] = nDens * sum(mgXSs * mgFlux)
-
-    return rxnRates
